+++ src/components/EcashTapButton.tsx
@@ -1,11 +1,11 @@
-import { useToast } from '@/hooks/useToast';
+import { useToast } from '@/hooks/util/useToast';
 import { RootState, useAppDispatch } from '@/redux/store';
 import { Modal } from 'flowbite-react';
 import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import SendEcashModalBody from './modals/SendEcashModalBody';
 import { resetStatus, setSending } from '@/redux/slices/ActivitySlice';
-import { useCashu } from '@/hooks/useCashu';
+import { useCashu } from '@/hooks/cashu/useCashu';
 
 export const BanknoteIcon = ({ className }: { className?: string }) => (
    <svg
