+++ src/components/buttons/lightning/Receive.tsx
@@ -4,7 +4,7 @@ import { Button, Modal, Spinner, Tooltip } from 'flowbite-react';
 import { ArrowDownRightIcon } from '@heroicons/react/20/solid';
 import { useDispatch, useSelector } from 'react-redux';
 import { setSuccess } from '@/redux/slices/ActivitySlice';
-import { useToast } from '@/hooks/useToast';
+import { useToast } from '@/hooks/util/useToast';
 import { assembleLightningAddress } from '@/utils/lud16';
 import ClipboardButton from '../utility/ClipboardButton';
 import QRCode from 'qrcode.react';
@@ -13,7 +13,7 @@ import ConfirmEcashReceiveModal from '@/components/modals/ConfirmEcashReceiveMod
 import { Token } from '@cashu/cashu-ts';
 import QRScannerButton from '../QRScannerButton';
 import { TxStatus, addTransaction, updateTransactionStatus } from '@/redux/slices/HistorySlice';
-import { useCashu } from '@/hooks/useCashu';
+import { useCashu } from '@/hooks/cashu/useCashu';
 
 const Receive = () => {
    const [isReceiveModalOpen, setIsReceiveModalOpen] = useState(false);
