+++ src/components/sidebar/Connections/NwcSidebarItem.tsx
