+++ src/components/sidebar/Taps/EcashTapsSettings.tsx
